+++ doc/CHANGELOG.md
@@ -37,6 +37,8 @@ The full set of tests is only run for tagged releases.
     - sys_tst_mig_n4d: low level MIG interface test
     - sys_tst_sram_n4d: memory test
     - sys_w11a_n4d: w11a with full 3840 MB memory
+  - for Digilent Arty S7
+    - sys_w11a_as7: w11a with full 3840 MB memory
 - new components
   - s7_cmt_sfs_2: dual-channel frequency synthesis MMCM/PLL wrapper
   - s7_cmt_1ce1ce2c: clocking block for 7-Series: 2 clk+CEs + 2 clk